Moved aggregate to separate package.
