+++ projects-management/src/main/java/pl/edu/knbit/organizer/teamrecruitment/MemberId.java
@@ -1,4 +1,4 @@
-package pl.edu.knbit.organizer;
+package pl.edu.knbit.organizer.teamrecruitment;
 
 public class MemberId {
