+++ projects-management/src/main/java/pl/edu/knbit/organizer/teamrecruitment/TeamRecruitmentFactory.java
@@ -1,4 +1,4 @@
-package pl.edu.knbit.organizer;
+package pl.edu.knbit.organizer.teamrecruitment;
 
 public class TeamRecruitmentFactory {
