+++ projects-management/src/main/java/pl/edu/knbit/organizer/teamrecruitment/TeamRecruitmentStatus.java
@@ -1,4 +1,4 @@
-package pl.edu.knbit.organizer;
+package pl.edu.knbit.organizer.teamrecruitment;
 
 public enum TeamRecruitmentStatus {
     OPEN, CLOSED, FINISHED;
