+++ projects-management/src/main/java/pl/edu/knbit/organizer/teamrecruitment/commands/FinishTeamRecruitmentCommand.java
@@ -1,14 +1,14 @@
-package pl.edu.knbit.organizer.commands;
+package pl.edu.knbit.organizer.teamrecruitment.commands;
 
 import org.axonframework.commandhandling.annotation.TargetAggregateIdentifier;
-import pl.edu.knbit.organizer.TeamRecruitmentId;
+import pl.edu.knbit.organizer.teamrecruitment.TeamRecruitmentId;
 
-public class CloseTeamRecruitmentCommand {
+public class FinishTeamRecruitmentCommand {
 
     @TargetAggregateIdentifier
     private final TeamRecruitmentId teamRecruitmentId;
 
-    public CloseTeamRecruitmentCommand(TeamRecruitmentId teamRecruitmentId) {
+    public FinishTeamRecruitmentCommand(TeamRecruitmentId teamRecruitmentId) {
         this.teamRecruitmentId = teamRecruitmentId;
     }
