+++ projects-management/src/main/java/pl/edu/knbit/organizer/teamrecruitment/events/MemberAppliedEvent.java
@@ -1,7 +1,7 @@
-package pl.edu.knbit.organizer.events;
+package pl.edu.knbit.organizer.teamrecruitment.events;
 
-import pl.edu.knbit.organizer.MemberId;
-import pl.edu.knbit.organizer.TeamRecruitmentId;
+import pl.edu.knbit.organizer.teamrecruitment.MemberId;
+import pl.edu.knbit.organizer.teamrecruitment.TeamRecruitmentId;
 
 public class MemberAppliedEvent {
