+++ projects-management/src/main/java/pl/edu/knbit/organizer/teamrecruitment/events/MemberRejectedEvent.java
@@ -1,15 +1,15 @@
-package pl.edu.knbit.organizer.events;
+package pl.edu.knbit.organizer.teamrecruitment.events;
 
-import pl.edu.knbit.organizer.MemberId;
-import pl.edu.knbit.organizer.TeamRecruitmentId;
+import pl.edu.knbit.organizer.teamrecruitment.MemberId;
+import pl.edu.knbit.organizer.teamrecruitment.TeamRecruitmentId;
 
-public class MeetingAppointedEvent {
+public class MemberRejectedEvent {
 
     private final TeamRecruitmentId teamRecruitmentId;
 
     private final MemberId memberId;
 
-    public MeetingAppointedEvent(TeamRecruitmentId teamRecruitmentId, MemberId memberId) {
+    public MemberRejectedEvent(TeamRecruitmentId teamRecruitmentId, MemberId memberId) {
         this.teamRecruitmentId = teamRecruitmentId;
         this.memberId = memberId;
     }
@@ -21,4 +21,5 @@ public class MeetingAppointedEvent {
     public MemberId getMemberId() {
         return memberId;
     }
+
 }
