+++ projects-management/src/test/java/pl/edu/knbit/organizer/TeamRecruitmentTest.java
@@ -4,9 +4,13 @@ import org.axonframework.test.FixtureConfiguration;
 import org.axonframework.test.Fixtures;
 import org.junit.Before;
 import org.junit.Test;
-import pl.edu.knbit.organizer.commands.ApplyForProjectCommand;
-import pl.edu.knbit.organizer.events.MemberAppliedEvent;
-import pl.edu.knbit.organizer.events.TeamRecruitmentOpenedEvent;
+import pl.edu.knbit.organizer.teamrecruitment.commands.ApplyForProjectCommand;
+import pl.edu.knbit.organizer.teamrecruitment.events.MemberAppliedEvent;
+import pl.edu.knbit.organizer.teamrecruitment.events.TeamRecruitmentOpenedEvent;
+import pl.edu.knbit.organizer.teamrecruitment.MemberId;
+import pl.edu.knbit.organizer.teamrecruitment.TeamRecruitment;
+import pl.edu.knbit.organizer.teamrecruitment.TeamRecruitmentCommandHandler;
+import pl.edu.knbit.organizer.teamrecruitment.TeamRecruitmentId;
 
 public class TeamRecruitmentTest {
